requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.9
django-allauth==0.49.0
django-cors-headers==3.4.0
django-heroku==0.3.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
djoser==2.0.3
gunicorn==20.1.0
idna==3.3
oauthlib==3.2.0
psycopg2==2.9.3
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-decouple==3.6
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
whitenoise==6.0.0
